6-spring-dao-orm v2. Resolved n+1 problem.

+++ 6-sping-dao-orm/src/main/java/ru/otus/spring/repository/AuthorRepository.java
@@ -5,5 +5,5 @@ import ru.otus.spring.domain.Author;
 import java.util.Optional;
 
 public interface AuthorRepository {
-    Optional<Author> findById(int id);
+    Optional<Author> findById(long id);
 }
