+++ 6-sping-dao-orm/src/main/java/ru/otus/spring/repository/BookCommentRepository.java
@@ -8,11 +8,11 @@ import java.util.Optional;
 public interface BookCommentRepository {
     BookComment save(BookComment comment);
 
-    Optional<BookComment> findById(int id);
+    Optional<BookComment> findById(long id);
 
     List<BookComment> findAll();
 
-    void updateTextById(int id, String text);
+    void updateTextById(long id, String text);
 
-    void deleteById(int id);
+    void deleteById(long id);
 }
