+++ 6-sping-dao-orm/src/main/java/ru/otus/spring/repository/GenreRepository.java
@@ -5,5 +5,5 @@ import ru.otus.spring.domain.Genre;
 import java.util.Optional;
 
 public interface GenreRepository {
-    Optional<Genre> findById(int id);
+    Optional<Genre> findById(long id);
 }
